Don't automatically chown unknown mountpoints to root:wheel in the event the user/group cannot be resolved on the FreeNAS box.
This can create undesirable side effects (in particular with NFS) if the
FreeNAS and client instance user/group list are out of synch and someone
logged into the box and changed the user/group to an 'unknown'
user/group (chown(2) allows numeric assignment anyhow -- for better or for
worse -- and chown(1) allows numeric / name based assignment).

This should be handled via Exceptions in the future, but for now
defaulting to root:wheel in the GUI alone is good enough.

Document the behavior accordingly.


Former-commit-id: ec5f5b438b8bf13942eccd0f942ed90215d3b208