Unbreak afpd configuration reload.
Unfortunately Warner did something right and I broke the reloading the configuration because I didn't realize that notifier HUPed afpd with its PID.

We can get away with killall -1 afpd as it doesn't blow away existing share connections.

The hammer solution of using restart for everything but AD / LDAP in the services toggle form really needs to be fixed though because it emits a ton more noise than necessary and almost doubles the service setup and teardown time unnecessarily, which makes FreeNAS look braindead compared with other solutions.

At least this is a first step in the right direction...


Former-commit-id: 648ac2cc6a420b06e737bf2467838680f489c214