Replication uses ssh to communicate data across the wire. We should ensure that both endpoints are present and talking when doing the replication, otherwise one side or the other could hang up, resulting in zombie connections, which could get our replication code into an odd funk where it wouldn't properly terminate, and thus could back up the system with multiple requests.
This should resolve some issues seen by a customer (Todd H.) as well as potentially other FreeNAS users.. but this is just a precautionary measure. The zfs send and recieve protocol may need to be cleaned up so that both ends check on each other periodically to ensure that they're in a semi-sane state. This may be a bug as well with ZFS v15 that has been resolved in later versions as well.

Lots of unknowns..


Former-commit-id: 5449ff39b7b12dd10d53e80af12f6ffe0775076b