Restore status quo for image naming scheme.
The problem after the build infrastructure was refactored by your's truly was that I didn't fully understand the importance of the path labeling for the full and GUI images.

So, let's restore the prefixing according to what's posted up on the FreeNAS SF site as the nuance of each image with the current naming scheme and previous naming scheme aren't incredibly intuitive.

Follow ${NANO_DISKIMGDIR} appropriate to create_*_diskimage, instead of ${NANO_OBJ} to more tightly couple the two functions.

Also, fix a legacy bug where nanobsd.sh -i would fail because it would always assume that the source file for the GUI upgrade image would always exist.

Clean up the ISO temporary directory so stale context doesn't accumulate in the 'temporary' directory (in reality this should be a legitimate temporary directory).

Don't regenerate the full install image in create_iso.sh. This is incredibly wasteful (both of CPU cycles and disk space): use a hardlink instead.


Former-commit-id: 983e0e1d75a2cc373ff5cbc9ca368e4dcd5ef4c7