Add a knob to enable/disable avahi from within the CIFS global settings, labeled as "Zeroconf discovery" (it's the technical term for what Bonjour and Avahi do), and add in needed bit to take away cifs advertisement when needed.
This solves the low hanging fruit with service interoperability, but like I discovered through testing, there are a number of gaps (still) where avahi isn't being reloaded or started properly (in part because of how we order / invoke the services and in part because of bugs in the rc.d scripts for avahi-app). You have to start/stop services as changing global settings doesn't actually reload configuration in an intuitive manner; I have a private patch that I believe will alleviate this, but it needs more testing before it's released into the wild.

This fixes the immediate issue for the item I requested in ticket # 970. Please note that while I requested that Avahi should be disabled by default, I was thinking of a one customer scenario where he wanted it off, when in reality home users (and other customers) most likely want out of the box configuration that works with Macs. So the needs of the many outweigh the needs of the few, thus Zeroconf discovery will remain on by default for now.


Former-commit-id: 66bcd7b9d6c2ade012b87ce0b9a70614ed35b562