Remove the try-except swallowing around volume 'exporting'.
try-except swallowing gets our software out of synch with reality.
It is also a really bad practice as it hides other potential bugs
with the code, and ultimately we should be fixing the issues, not
disguising bugs in our code.

Doing this also forces the sysadmin to deal with the issue ASAP
instead of causing head-scratching later on when things aren't
working as expected, as I've had to deal with over the
phone/IRC/etc with several users/customers.


Former-commit-id: aa61f31e8551dc47957bc4bd299bb9289e40b90c