Clean up cornercase where a zvol, dataset, zpool in use when deleting a zvol would delete the zvol from the GUI/db, but wouldn't actually remove things from the backend in FreeBSD.
Introduce a few new notifier APIs to help ease this pain:

__get_mountpath - get the mountpath for a volume/dataset (name)
__rmdir_mountpoint - remove a mountpoint as a means to determine whether or not the

Split storage.models.Volumes.delete into two functions again to 'clean up' the database after storage.models.Volumes._delete has been run, in such a manner that the mountpoints aren't preemptively removed, even though some of the dataset / zvol objects might have been successfully removed, and instead are removed in the event that the directory, that previously was the mountpoint, is no longer present.

These APIs are coded relatively conservatively for now to avoid putting the user into situations where they need to do database surgery.. they can and should be cleaned up later to be more explicit and straightforward by tossing proper Exceptions up the stack.

This doesn't catch the following cases:

mount -f vol

zpool destroy -f tank

when I'm in /mnt/vol or /mnt/tank (this occurs on 8.0.3 as well and this is sort of by design as they're 'force' operations). This might be a noop, or something of concern.

This change -- along with r9317 and r9318 -- resolves the issue where one couldn't export volumes via the GUI, as explained in ticket # 1126.


Former-commit-id: a487cd5122c4974912c7c646009376f210601ceb