Use /var/home indirection instead of manipulating /home for AD/LDAP.
This change adds a single layer of indirection from /home -> /var/home
in the install image such that we no longer have to touch /home in the
when home directories are specified in the CIFS configuration when
ix-activedirectory and ix-ldap are run.

Verified sane via the AD code path with both trusted domains being set
and not being set. Please note that the ix-ldap code path is a simpler
pattern than the one found in ix-activedirectory, and thus covering
ix-activedirectory is sufficient.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>

Former-commit-id: 5178999021cbda62dd5249c4687bf56d285b286d