Donate autotuner app from TrueNAS to FreeNAS project.
The defaults passed with ix-loader to autotune.py might need to
be tweaked in the future, depending on the number of PBIs,
features added, etc. The script is based on some observations
made when dealing with some TrueNAS customers in the past.

The script will need to be modified to put entries into
the database so they can be manipulated by end-users in the
GUI.

NOTE: do not modify /boot/loader.conf or your machine will
reboot by design to adjust the values that it sets.

There will be a reboot to apply the change, but after that the
values will persist in /boot/loader.conf until upgraded (in
which case the autotune script will be run via ix-loader in
ix-update, again by design).

Sponsored-by: iXsystems, Inc.
Requested-by: matto, jpaetzel

Do-Not-Merge (ix): already in the ix repo.


Former-commit-id: c21f10261844c394e378fdfcae62a17bb32f6a9f