Strip binaries more intelligently
We really only care about stripping binaries in bin/ and sbin/
and .so files in lib/ . .a files get nuked, .la files are just
text and there are a load of other binary blobs hanging out in
usr/local/lib .

The FreeBSD base build system already does a proper job in bin/,
sbin/, etc, so that's not in question.

The ports system could handle this but it doesn't 100%. Also,
switching from DEBUG=0 and DEBUG=1 would require rebuilding all
packages (waste of time and effort).

Also, the problem with the previous method of stripping
(SF @ r8726) is that some packages like installing binaries in
locations that (IMHO) really don't make sense. Example,
open-vm-tools places kernel modules in /usr/local/lib/

This does the following:
1. Fixes open-vm-tools on branches/8.2.0 (-> DEBUG not specified
   on SW_FEATURES) and subsequently fuse-kmod if it was still
   installed in /usr/local .
2. Should speed up the build/hit the build server's disk less.


Former-commit-id: a195f9bf6553a067c2b79c36b04f26f7afbdf5ff