Fixup boottime, add passthrough json, prompt user.
Change the --boot option to only send the crash dump.  Also
switch the --boot option to be '-b'.  This is so we can only
send the non-sensitive information at boot time.

Add a prompt to allow the user to opt out about sending logs
and freenas-debug to us if they choose.  Add a -F flag to supress
this warning when run from the GUI or other scripts.

Allow -t to pass through a file into the crashuploader script,
this file will be put into the crash dir as "ticket_info.json"
and will be passed through to the crashuploader script.


Former-commit-id: d9455141b623d1d004059d8983f8789fe2c27d78