Fix issues in 738f085e a298b3be.
create_iso.sh generated the 0005.verify_media_size.sh script
and in the process would write the current value (during build)
of IMG_SIZE hardwired into the script.

In addition most of the other variables were escaped with \$, however
now that the file is being copied the variables no longer need to
be escaped.


Former-commit-id: 59bf893f5660fe8d2c650ec80be19860e4a8bdad