The default mkisofs command created a volume label of "CDROM". Change it to create a volume label of "FreeNAS_INSTALL". During boot of the CDROM, geom will create a device entry /dev/iso9660/FreeNAS_INSTALL. This can be mounted via:
mount -t cd9660 /dev/iso9660/FreeNAS_INSTALL /mnt

In /boot/loader.conf for the CDROM, set vfs.root.mountfrom to "cd9660:iso9660/FreeNAS_INSTALL"
No matter what underlying device the CDROM is on, this will allow the installer
to find it and mount it.

This allows the FreeNAS CDROM to be booted under BHyve, which seems to boot
a CD-ROM on /dev/vtbd1.


Former-commit-id: 1e7d0438911c9d880d99d4c723fd9fe0435300f3