When the build fails, print out the last log file that was written to.
The nanobsd.sh script used in the FreeNAS build writes to
many small log files.  If the build fails, these small
log files are not displayed, and the the user is instructed
to "please check above log for more details".
Improve the nanobsd.sh script slightly, so that it keeps track
of the last log file written to, and if it exits due to an
error, display this log file.  This helps slightly when
using Jenkins Continuous Integration for doing builds and detecting
build breaks.


Former-commit-id: 852660c0ebcd54f5d7f318727b53ed23a2ebd504