Use the django ORM to generate the config file for netatalk.
This is currently completely broken, as the startup script simply
calls a python script that then uses the ORM to grab a few things,
but does NOT generate a valid config file for netatalk2 or netatalk3.

In addition because the ORM is so expensive we are going to need
a way to not call it if no config needs to be generated.  So the
rc.d script will need to be smarter.


Former-commit-id: e7acef7df9c43fba4d0f79e672a4ec6153d7bf9a