Move add_pbi_manager() before add_gui() in the list of build steps specified by customize_cmd.
add_pbi_manager() installs tools such as /sbin/pbi_listrepo in the chroot used
during the generation of the GUI.  In add_gui() the Django "python manage.py syncdb"
command is called which triggers the South migration.  The South migration
in turn may try to invoke /sbin/pbi_listrepo.  On the Jenkins build host,
there seemed to be a race condition triggered where add_gui() was called,
but /sbin/pbi_listrepo did not exist.  This resulted in a build break and
Python stack trace during the build.  This should fix that.


Former-commit-id: fef1dcd01dde8f3cff75f421965d1dca24bd0496