Just generate /etc/exports
The middleware does all of the manipulation of services when shares are
added or services are started or modified.  The block of code this commit
removes causes mountd and rpcbind to start unconditionally at boot, regardless
od whether the service is enabled or not.  The code that kicks mountd
after regenerating /etc/exports is removed as well, but the current middleware
handles reloading mountd in the reload_nfs method.

Historically the middleware code wasn't as smart and the rc.d scripts needed
to have more intelligence in them.  I believe this is leftover from those times.
When the middleware was improved this rc.d script was left unchanged.


Former-commit-id: 825c1c5012c2abab32b991b9927691951dcef6a4