Do not rely on "make package" doing a 'make install'.
With the new FreeBSD ports "staging" framework it appears
that packages do not need to be installed for packages to
be built.  Since we assumed that
"make package" = "make install package"
we were winding up with ports not installed when doing a fresh
build due to packages being built however not installed.

Later builds would actually pkg_add the packages giving false
build success positives.


Former-commit-id: e9d366f9a4eb16b47510ee19dc083dc375156ad8