Move AFP CNID back to the share directory.
In netatalk 2.x the CNID was contained in the share directory itself.
In netatalk 3.0 this moved to /var/netatalk.

This is a problem on FreeNAS with a large share and it's small /var
After updating netatalk to 3.1.0 a boolean was added that restores
the Netatalk 2.x behavior of storing the CNID in the share.


Former-commit-id: ac8b9e168ed637f4e46d13102ec428c164685733