Improve the clean targets
Fix flaws in the script called by make clean in an
ongoing effort to have make clean DTRT, which
is restore a build environment to it's virgin
state.

Instead of depending on packages being where they
should be just go ahead and search the entire workspace
for them and nuke as they are found.  I know there
are symlinks to the packages, so this isn't complete
paranoia.  However it is a lot slower.


Former-commit-id: 3a632ec71a31c954bc114271c821f796be5b81da