Fix to build releases with thin images.
The way nanobsd works, it first builds a thick image
which is used to compile all of the ports needed.  This
image has things like gcc, make a toolchain, etc.

A side effect of building ports to place in the image is
packages are created.  Subsequent builds using that build
environment will use the packages and a "thin" image is
created which does not have a toolchain.


Former-commit-id: 78d21da6d4e5ea4e62b6b6b040f7da737d776ae9