Skip over null bytes when dd'ing the image
The FreeNAS image has the space for slice 2 preallocated
but not filled with anything.  This slice is used for upgrades.

There's no reason to try and write the null bytes when installing
for the ISO.  Go ahead and skip over them. This pretty much
halves the install time, and restores accuracy to the percentage
complete display. (It no longer spins on 99% for half the install)


Former-commit-id: 77ad8878b8502aa1b135eabc806ed2b6062f6a3c