Make sure ix scripts run early enough
Since the ix scripts were moved from /etc/rc.d they were not running
early enough in the rc system because only /etc/rc.d are taken into
account for the early boot stage.

Segregating boot stage is not crucial for us because everything is in
the same filesystem or available at that time.


Former-commit-id: ec35a5efbe174a569ce4503949b49dec11b47c9e