Teardown and recreate vlan interfaces when networking is restarted.
Ticket: #3708

By doing this any and all changes to a vlan interface are immediately
done when the GUI is changed.  The previous behavior was open to
a number of bugs that could be caused by renaming the interface or
changing it's parent interface.  In the old behavior only a reboot
would ensure the correct configuration.


Former-commit-id: 8b261d4e9761db5b8c6c09ba9dfe9ee67aa02711