Pump the zfs send pipe on our own
Python subprocess.Popen doesn't handle large stream of data very well
between process in communicate(), it was causing zfs send to become a
zombie, hanging the second part of the shell.

While we are here we can get rid the dd in the pipe as we are doing it
in the loop already.

Ticket:	#4600

Former-commit-id: d27ff8fd0d87552343bb8508417d4c065c975a46