For ports which want openldap, we need to set the WANT_OPENLDAP_SASL option in order to properly record the dependency between the ports using openldap and the openldap-sasl port.
In the non-poudriere build of ports, things seemed to work
because we built openldap-sasl early in the last of ports.

In the poudriere build of ports, this dependency was not recorded
properly so some ports were built against the openldap port, and some
were built against the openldap-sasl port.  pkg_add then failed
to install all the ports due to dependency conflicts,
because openldap and openldap-sasl ports were installed as dependent ports.

This fix records the depencies properly, so all the ports build properly
with poudriere, and install without conflicts.


Former-commit-id: 68cccefd9582f256af529e8fb807f0b99107d240