If there is a system dataset, use it.
Ticket: #5124

The issue here is that if there's a system dataset that is being used for syslog
the logfiles could grow to be quite large.  When ixdiagnose would try to
make it's tempfiles it would make them on /var, which is a memory
filesystem.

This would cause ixdiagnose to fail.


Former-commit-id: 9593467853fea7c10693aaf23b1155ce88bd5a53