Remove cuteness from code.
Ticket: #5267

While I would hesitate to ADD code to a release this late in the game, removing
code...what could possibly go wrong?

This commit removes the ability for you to pass in either a UID or a username
to the function that resolves users.  It also removes the ability to pass
in a GID or groupname to the code that resolves groups.

I'll note that NOWHERE in the code is a UID or GID passed in to this function.

Upon seeing an integer the code assumes it's a GID/UID.

Which causes fail because what the GUI passes in is a username/groupname.

So if the username/groupname is a number, then the code tries to resolve the GUI/UID,
and if it just so happens that the "name" you use doesn't match an existing GID/UID,
then a NoneType gets returned as the resolution fails, which eventually gets passed
back to the GUI and results in a traceback.

In reality I have grepped the code, introspected it, and talked to John, and now
I'm mostly hoping that if my commit message is long enough people will just let
me commit this to 9.2.1-BRANCH in the hopes that I'll just fucking stop.

However if we don't commit this then we really ought to add code in the form
of a validator to the GUI which blocks adding users or groups with numerical names,
either that or we'll be shipping 9.2.1.6 with a KNOWN BUG; a local DoS for root users,
which is an unforgivable and egregious sin which will cause me to doubt my ability as
a software engineer.  Given the choice between adding code, removing code or shipping
with bugs I'd prefer to remove code, but hey, you do what you want man!


Former-commit-id: 2a9480434451ff23d820d37992c4779e6a24a4aa